+++ src/main/java/com/solvd/demoapp/pages/common/SauceLabsBase.java
@@ -1,18 +1,17 @@
 package com.solvd.demoapp.pages.common;
 
+import com.solvd.demoapp.pages.ios.SauceLabs;
 import com.zebrunner.carina.utils.ios.IOSUtils;
-import com.zebrunner.carina.webdriver.decorator.ExtendedWebElement;
 import com.zebrunner.carina.webdriver.gui.AbstractPage;
-import com.zebrunner.carina.webdriver.locator.ExtendedFindBy;
 import org.openqa.selenium.WebDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public abstract class SauceLabsBase extends AbstractPage implements IOSUtils {
-
-    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeImage[`name == \"Saucelabs\"`]")
-    private ExtendedWebElement logo;
+    private static final Logger LOGGER = LoggerFactory.getLogger(SauceLabs.class);
 
     public SauceLabsBase(WebDriver driver) {
         super(driver);
-        setUiLoadedMarker(logo);
+        LOGGER.info("SauceLabsBase()");
     }
 }
