--- src/main/java/com/solvd/demoapp/pages/ios/RightMenuPage.java
@@ -1,17 +1,0 @@
-package com.solvd.demoapp.pages.ios;
-
-import com.solvd.demoapp.pages.common.RightMenuPageBase;
-import com.zebrunner.carina.utils.factory.DeviceType;
-import org.openqa.selenium.WebDriver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@DeviceType(pageType = DeviceType.Type.IOS_PHONE, parentClass = RightMenuPageBase.class)
-public class RightMenuPage extends RightMenuPageBase {
-    private static final Logger LOGGER = LoggerFactory.getLogger(RightMenuPage.class);
-
-    public RightMenuPage(WebDriver driver) {
-        super(driver);
-        LOGGER.info("RightMenuPage()");
-    }
-}
