+++ src/main/resources/_config-android.properties
@@ -1,3 +1,7 @@
+capabilities.appPackage=com.saucelabs.mydemoapp.android
+capabilities.bundleId=com.saucelabs.mydemoapp.android
+capabilities.appActivity=com.saucelabs.mydemoapp.android.view.activities.SplashActivity
+
 capabilities.automationName=UiAutomator2
 capabilities.platformName=Android
 capabilities.udid=emulator-5554
