+++ src/main/resources/_config-ios.properties
@@ -1,3 +1,6 @@
+capabilities.appPackage=com.saucelabs.mydemoapp.ios
+capabilities.bundleId=com.saucelabs.mydemoapp.ios
+
 capabilities.platformName=iOS
 capabilities.deviceName=iPhone 15 Pro Max
 capabilities.deviceType=phone
