+++ src/main/resources/_config.properties
@@ -1,6 +1,3 @@
-capabilities.appPackage=com.saucelabs.mydemoapp.ios
-capabilities.bundleId=com.saucelabs.mydemoapp.ios
-
 page_opening_strategy=BY_ELEMENT
 element_loading_strategy=BY_VISIBILITY
