+++ src/main/resources/_testdata.properties
@@ -1,5 +1,5 @@
 username-demo=bob@example.com
 password-demo=10203040
 
-location-longitude=13.45143
-location-latitude=52.50032
+location-longitude=-122.406417
+location-latitude=37.785834
